BF: use explicit use_directory_urls: true and provide site_url
We had used those urls which end with / not .html, and presumably
something changed on mkdocs side (or who knows) and now it started
to build with .html. To enable use_directory_urls: true we also had to
provide site_url which I do here.

Commit is done under careful guidance by Satra